Workaround retrieving tokens from broken providers
Some OAuth2 providers such as Google or Github do not follow the OAuth2
spec to retrieve the access token. The package currently works around
this by maintaining a short interal whitelist of non-compliant OAuth
providers.

Maintaining an internal whitelist means some big non-compliant
providers cannot be used unless they are added to the whitelist by
either forking this package or trying to get it added to the whitelist.

This commit proposes being able to setup the config to build a custom
URL following the non-standard requirements. It uses a field name in
the config that clearly highlights setting it up means the provider is
broken and non-compliant with the spec.